A more TypeScripty syntax
